Create server library
