+++ rust/grpc-api/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "grpc-hello-world"
+name = "grpc-api"
 version = "0.1.0"
 authors = ["Florian Gilcher <florian.gilcher@asquera.de>"]
 edition = "2018"
@@ -11,10 +11,10 @@ version = "0.5.0-alpha.4"
 
 [dependencies]
 xain-proto = { path = "../xain-proto"}
-futures = "*"
-slog = "*"
-slog-async = "*"
-slog-scope = "*"
-slog-stdlog = "*"
-slog-term = "*"
-log = "*"
+futures = "0.1"
+slog = "2"
+slog-async = "2"
+slog-scope = "2"
+slog-stdlog = "2"
+slog-term = "2"
+log = "0.4"
